Non-optimized finalized
That is, no optimization has been done. Also, all operations on points
within the same space, no between-space operations (compatibility of
abc sequences and so on).

Also, more fundamentally, errors are thrown as exceptions. Using
something like an Option might be preferable but it could also scare
people away which is not the goal here.